Moved all flood fill implementation to Pathing.js
