+++ Slime_game/classes/Pathing.js
@@ -1,3 +1,4 @@
+import {scene} from "./Controller.js";
 //board: the array Board that holds Tile objects
 //start: a list holding the [x,y] coordinates of the start position
 //destination: a list holding the [x,y] coordinates of the end position
@@ -49,3 +50,54 @@ function getNeighbors(board, currentNode, currentHeight, maxHeight) {
         }
     }
 }
+
+//FLOOD FILL IMPLEMENTATION
+function hover(board){//initiates methods when cursor hovers over entities/tiles
+    var cPos = board.cursor.position;
+    var type = board.tileMap[cPos[0]][cPos[2]];
+    var height = board.heightMap[cPos[0]][cPos[2]];
+    var pPos = board.player.position;
+    //board.cursor.cursorHeight(height + 0.5);
+    console.log("Type: ", typeList(type));
+    console.log("Height: ", height);
+    console.log("Occupied by: ", occupied(board.player, pPos, cPos));
+    //board.cursor.position = (cPos[0], height + 0.5, cPos[2]);
+    //return type;
+}
+
+function typeList(type){//Returns the terrain name for logging to console
+    switch(type){
+        
+        case 0:
+            return "Grass";
+        case 1:
+            return "Rocky";
+        case 2:
+            return "Water";
+        case 3:
+            return "Gap";
+        case 4:
+            return "Cave";
+        case 8:
+            return "Exit";
+        case 9://may change as board gen changes
+            return "Void";
+    }
+}
+
+function occupied(player, pPos, cPos){
+    //var occupied = false;
+    if(pPos[0] == cPos[0] && pPos[2] == cPos[2]){
+        //occupied == true;
+        var overlayList = player.movementOverlayHelper(pPos[0], pPos[2], player.movementRange, 1);//will need to read player height in future
+        for(var i = 0; i < overlayList.length; i++){
+            scene.add(overlayList[i]);
+        }
+        return "Player";
+    }
+    else{
+        return "None";
+    }
+}
+
+export {hover};
